refactor: implement import symbols from additional file js
